# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AoraGrand Hotel Booking API",
        "description": "RESTful API for AoraGrand Hotel Booking System with Midtrans payment integration",
        "contact": {
            "name": "AoraGrand Hotel Support",
            "email": "admin@aoragrand.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        },
        {
            "url": "https://api.aoragrand.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/room-types": {
            "get": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Get all room types",
                "operationId": "ca013b8668b9235311faca0f7063a058",
                "parameters": [
                    {
                        "name": "check_in_date",
                        "in": "query",
                        "description": "Check-in date to check availability",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-12-01"
                        }
                    },
                    {
                        "name": "check_out_date",
                        "in": "query",
                        "description": "Check-out date to check availability",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-12-03"
                        }
                    },
                    {
                        "name": "guests",
                        "in": "query",
                        "description": "Number of guests",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room types retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Room types retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RoomType"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/room-types": {
            "post": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Create a new room type (Admin only)",
                "operationId": "321fedf14e9293e2e80f625fcf73eae5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price_per_night",
                                    "capacity",
                                    "total_rooms"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Standard Room"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Comfortable standard room"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "total_rooms": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "facilities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "AC",
                                            "TV",
                                            "WiFi"
                                        ]
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "/images/room1.jpg"
                                        ]
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Room type created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoomType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/room-types/{id}": {
            "get": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Get a specific room type",
                "operationId": "32dfe7550e8eb4648f15f4037cca60ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "check_in_date",
                        "in": "query",
                        "description": "Check-in date to check availability",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-12-01"
                        }
                    },
                    {
                        "name": "check_out_date",
                        "in": "query",
                        "description": "Check-out date to check availability",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-12-03"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room type retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Room type retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoomType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                }
            }
        },
        "/api/admin/room-types/{id}": {
            "put": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Update a room type (Admin only)",
                "operationId": "41ac828c3380f2cfd9e8a5a995b09da3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Standard Room"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Comfortable standard room"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "total_rooms": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "facilities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "AC",
                                            "TV",
                                            "WiFi"
                                        ]
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "/images/room1.jpg"
                                        ]
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room type updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Room type updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoomType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Admin access required"
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Delete a room type (Admin only)",
                "operationId": "a078198d59db1b0a5926df1bd3dc5bf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room type deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Room type deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Admin access required"
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "RoomType": {
                "title": "Room Type",
                "description": "Room Type model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Standard Room"
                    },
                    "description": {
                        "type": "string",
                        "example": "Comfortable standard room with basic amenities"
                    },
                    "price_per_night": {
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_rooms": {
                        "type": "integer",
                        "example": 20
                    },
                    "available_rooms": {
                        "type": "integer",
                        "example": 15
                    },
                    "facilities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "AC",
                            "TV",
                            "WiFi"
                        ]
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "/images/room1.jpg",
                            "/images/room2.jpg"
                        ]
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}